TST: implement test_formats for Index
